MindMetricAI/backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
scikit-learn
pandas
numpy
openai
google-generativeai
